requirements.txt
transformers
datasets
torch
torchvision
diffusers
numpy
pandas
scikit-learn
opencv-python
tqdm
argparse
json5
Pillow
accelerate
xformers
